implement nested list with marsh
